working version

+++ src/main/java/com/springrest/springrest/services/interfaces/ICourseService.java
@@ -7,10 +7,10 @@ import java.util.List;
 public interface ICourseService {
     public List<Course> getCourses();
 
-    public Course getCourse(int courseId);
+    public Course getCourse(long courseId);
 
     public Course addCourse(Course course);
 
     public Course updateCourse(Course course);
-    public void deleteCourse(int courseId);
+    public void deleteCourse(long courseId);
 }
